DataFrame-class {SparkR}	R Documentation

S4 class that represents a DataFrame

Description

DataFrames can be created using functions like jsonFile, table etc.

Compute aggregates by specifying a list of columns

Usage

dataFrame(sdf, isCached = FALSE)

## S4 method for signature 'DataFrame'
agg(x, ...)

## S4 method for signature 'DataFrame'
summarize(x, ...)

groupBy(x, ...)

groupedData(sgd)

Arguments

sdf
A Java object reference to the backing Scala DataFrame

isCached
TRUE if the dataFrame is cached

x
a DataFrame

env
An R environment that stores bookkeeping states of the DataFrame

sdf
A Java object reference to the backing Scala DataFrame

See Also

jsonFile, table

[Package SparkR version 1.4.0 Index]