RDD-class {SparkR}	R Documentation

S4 class that represents an RDD

Description

RDD can be created using functions like parallelize, textFile etc.

Usage

RDD(jrdd, serializedMode = "byte", isCached = FALSE,
  isCheckpointed = FALSE)

Arguments

jrdd
Java object reference to the backing JavaRDD

serializedMode
Use "byte" if the RDD stores data serialized in R, "string" if the RDD stores strings

isCached
TRUE if the RDD is cached

isCheckpointed
TRUE if the RDD has been checkpointed

Slots

env

An R environment that stores bookkeeping states of the RDD

jrdd

Java object reference to the backing JavaRDD

See Also

parallelize, textFile

[Package SparkR version 0.1 Index]