attach,DataFrame-method {SparkR}	R Documentation

Attach DataFrame to R search path

Description

The specified DataFrame is attached to the R search path. This means that the DataFrame is searched by R when evaluating a variable, so columns in the DataFrame can be accessed by simply giving their names.

Usage

## S4 method for signature 'DataFrame'
attach(what, pos = 2,
  name = deparse(substitute(what)), warn.conflicts = TRUE)

attach(what, pos = 2L, name = deparse(substitute(what)),
  warn.conflicts = TRUE)

Arguments

what
(DataFrame) The DataFrame to attach

pos
(integer) Specify position in search() where to attach.

name
(character) Name to use for the attached DataFrame. Names starting with package: are reserved for library.

warn.conflicts
(logical) If TRUE, warnings are printed about conflicts from attaching the database, unless that DataFrame contains an object

See Also

detach

Other DataFrame functions: $, $<-, select, select, select,DataFrame,Column-method, select,DataFrame,list-method, selectExpr; DataFrame-class, dataFrame, groupedData; [, [, [[, subset; agg, agg, count,GroupedData-method, summarize, summarize; arrange, arrange, arrange, orderBy, orderBy; as.data.frame, as.data.frame,DataFrame-method; cache; collect; colnames, colnames, colnames<-, colnames<-, columns, names, names<-; coltypes, coltypes, coltypes<-, coltypes<-; columns, dtypes, printSchema, schema, schema; count, nrow; describe, describe, describe, summary, summary, summary,PipelineModel-method; dim; distinct, unique; dropna, dropna, fillna, fillna, na.omit, na.omit; dtypes; except, except; explain, explain; filter, filter, where, where; first, first; groupBy, groupBy, group_by, group_by; head; insertInto, insertInto; intersect, intersect; isLocal, isLocal; join; limit, limit; merge, merge; mutate, mutate, transform; ncol; persist; printSchema; rbind, rbind, unionAll, unionAll; registerTempTable, registerTempTable; rename, rename, withColumnRenamed, withColumnRenamed; repartition; sample, sample, sample_frac, sample_frac; saveAsParquetFile, saveAsParquetFile, write.parquet, write.parquet; saveAsTable, saveAsTable; saveDF, saveDF, write.df, write.df; selectExpr; showDF, showDF; show, show, show,GroupedData-method; take; transform, withColumn, withColumn; unpersist; write.json, write.json

Examples

## Not run: 
##D attach(irisDf)
##D summary(Sepal_Width)
## End(Not run)

[Package SparkR version 1.6.0 Index]