awaitTermination {SparkR}	R Documentation

awaitTermination

Description

Waits for the termination of the query, either by stopQuery or by an error.

Usage

awaitTermination(x, timeout = NULL)

## S4 method for signature 'StreamingQuery'
awaitTermination(x, timeout = NULL)

Arguments

x
a StreamingQuery.

timeout
time to wait in milliseconds, if omitted, wait indefinitely until stopQuery is called or an error has occurred.

Details

If the query has terminated, then all subsequent calls to this method will return TRUE immediately.

Value

TRUE if query has terminated within the timeout period; nothing if timeout is not specified.

Note

awaitTermination(StreamingQuery) since 2.2.0

experimental

See Also

Other StreamingQuery methods: explain(), isActive(), lastProgress(), queryName(), status(), stopQuery()

Examples

## Not run:  awaitTermination(sq, 10000) 

[Package SparkR version 3.2.3 Index]