base64 {SparkR}	R Documentation

base64

Description

Computes the BASE64 encoding of a binary column and returns it as a string column. This is the reverse of unbase64.

Usage

base64(x)

## S4 method for signature 'Column'
base64(x)

Arguments

x
Column to compute on.

Note

base64 since 1.5.0

See Also

Other string_funcs: ascii, concat_ws, concat, decode, encode, format_number, format_string, initcap, instr, length, levenshtein, locate, lower, lpad, ltrim, regexp_extract, regexp_replace, reverse, rpad, rtrim, soundex, substring_index, translate, trim, unbase64, upper

Examples

## Not run: base64(df$c)

[Package SparkR version 2.1.2 Index]