setBroadcastValue {SparkR}	R Documentation

Internal function to set values of a broadcast variable.

Description

This function is used internally by Spark to set the value of a broadcast variable on workers. Not intended for use outside the package.

Usage

setBroadcastValue(bcastId, value)

Arguments

bcastId
The id of broadcast variable to set

value
The value to be set

See Also

broadcast, value

[Package SparkR version 0.1 Index]