cacheTable {SparkR}	R Documentation

Cache Table

Description

Caches the specified table in-memory.

Usage

## Default S3 method:
cacheTable(tableName)

Arguments

tableName
the qualified or unqualified name that designates a table. If no database identifier is provided, it refers to a table in the current database.

Value

SparkDataFrame

Note

cacheTable since 1.4.0

Examples

## Not run: 
##D sparkR.session()
##D path <- "path/to/file.json"
##D df <- read.json(path)
##D createOrReplaceTempView(df, "table")
##D cacheTable("table")
## End(Not run)

[Package SparkR version 2.4.1 Index]