column {SparkR}	R Documentation

S4 class that represents a SparkDataFrame column

Description

The column class supports unary, binary operations on SparkDataFrame columns

Returns a Column based on the given column name.

Usage

column(x)

## S4 method for signature 'jobj'
column(x)

## S4 method for signature 'character'
column(x)

Arguments

x
Character column name.

Slots

jc

reference to JVM SparkDataFrame column

Note

Column since 1.4.0

column since 1.6.0

See Also

Other non-aggregate functions: column_nonaggregate_functions, not()

Examples

## Not run: column("name")

[Package SparkR version 3.2.4 Index]