cov {SparkR}	R Documentation

cov

Description

Compute the covariance between two expressions.

Usage

cov(x, ...)

covar_samp(col1, col2)

covar_pop(col1, col2)

## S4 method for signature 'characterOrColumn'
cov(x, col2)

## S4 method for signature 'characterOrColumn,characterOrColumn'
covar_samp(col1, col2)

## S4 method for signature 'characterOrColumn,characterOrColumn'
covar_pop(col1, col2)

## S4 method for signature 'SparkDataFrame'
cov(x, colName1, colName2)

Arguments

x
a Column or a SparkDataFrame.

...
additional argument(s). If x is a Column, a Column should be provided. If x is a SparkDataFrame, two column names should be provided.

col1
the first Column.

col2
the second Column.

colName1
the name of the first column

colName2
the name of the second column

Details

cov: Compute the sample covariance between two expressions.

covar_sample: Alias for cov.

covar_pop: Computes the population covariance between two expressions.

cov: When applied to SparkDataFrame, this calculates the sample covariance of two numerical columns of one SparkDataFrame.

Value

The covariance of the two columns.

Note

cov since 1.6.0

covar_samp since 2.0.0

covar_pop since 2.0.0

cov since 1.6.0

See Also

Other aggregate functions: avg, column_aggregate_functions, corr, count, first, last

Other stat functions: approxQuantile, corr, crosstab, freqItems, sampleBy

Examples

## Not run: 
##D df <- createDataFrame(cbind(model = rownames(mtcars), mtcars))
##D head(select(df, cov(df$mpg, df$hp), cov("mpg", "hp"),
##D                 covar_samp(df$mpg, df$hp), covar_samp("mpg", "hp"),
##D                 covar_pop(df$mpg, df$hp), covar_pop("mpg", "hp")))
## End(Not run)

## Not run: 
##D cov(df, "mpg", "hp")
##D cov(df, df$mpg, df$hp)
## End(Not run)

[Package SparkR version 2.4.1 Index]