dense_rank {SparkR}	R Documentation

dense_rank

Description

Window function: returns the rank of rows within a window partition, without any gaps. The difference between rank and dense_rank is that dense_rank leaves no gaps in ranking sequence when there are ties. That is, if you were ranking a competition using dense_rank and had three people tie for second place, you would say that all three were in second place and that the next person came in third. Rank would give me sequential numbers, making the person that came in third place (after the ties) would register as coming in fifth.

Usage

dense_rank(x = "missing")

## S4 method for signature 'missing'
dense_rank()

Arguments

x
empty. Should be used with no argument.

Details

This is equivalent to the DENSE_RANK function in SQL.

Note

dense_rank since 1.6.0

See Also

Other window_funcs: cume_dist, lag, lead, ntile, percent_rank, rank, row_number

Examples

## Not run: 
##D   df <- createDataFrame(mtcars)
##D   ws <- orderBy(windowPartitionBy("am"), "hp")
##D   out <- select(df, over(dense_rank(), ws), df$hp, df$am)
## End(Not run)

[Package SparkR version 2.1.2 Index]