dropTempView {SparkR}	R Documentation

Drops the temporary view with the given view name in the catalog.

Description

Drops the temporary view with the given view name in the catalog. If the view has been cached before, then it will also be uncached.

Usage

dropTempView(viewName)

Arguments

viewName
the name of the temporary view to be dropped.

Value

TRUE if the view is dropped successfully, FALSE otherwise.

Note

since 2.0.0

Examples

## Not run: 
##D sparkR.session()
##D df <- read.df(path, "parquet")
##D createOrReplaceTempView(df, "table")
##D dropTempView("table")
## End(Not run)

[Package SparkR version 2.4.1 Index]