first {SparkR}	R Documentation

Return the first row of a SparkDataFrame

Description

Return the first row of a SparkDataFrame

Aggregate function: returns the first value in a group.

Usage

first(x, ...)

## S4 method for signature 'SparkDataFrame'
first(x)

## S4 method for signature 'characterOrColumn'
first(x, na.rm = FALSE)

Arguments

x
a SparkDataFrame or a column used in aggregation function.

...
further arguments to be passed to or from other methods.

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

Details

The function by default returns the first values it sees. It will return the first non-missing value it sees when na.rm is set to true. If all values are missing, then NA is returned.

Note

first(SparkDataFrame) since 1.4.0

first(characterOrColumn) since 1.4.0

See Also

Other SparkDataFrame functions: SparkDataFrame-class, agg, arrange, as.data.frame, attach, cache, coalesce, collect, colnames, coltypes, createOrReplaceTempView, crossJoin, dapplyCollect, dapply, describe, dim, distinct, dropDuplicates, dropna, drop, dtypes, except, explain, filter, gapplyCollect, gapply, getNumPartitions, group_by, head, histogram, insertInto, intersect, isLocal, join, limit, merge, mutate, ncol, nrow, persist, printSchema, randomSplit, rbind, registerTempTable, rename, repartition, sample, saveAsTable, schema, selectExpr, select, showDF, show, storageLevel, str, subset, take, union, unpersist, withColumn, with, write.df, write.jdbc, write.json, write.orc, write.parquet, write.text

Other agg_funcs: agg, avg, countDistinct, count, kurtosis, last, max, mean, min, sd, skewness, stddev_pop, stddev_samp, sumDistinct, sum, var_pop, var_samp, var

Examples

## Not run: 
##D sparkR.session()
##D path <- "path/to/file.json"
##D df <- read.json(path)
##D first(df)
## End(Not run)
## Not run: 
##D first(df$c)
##D first(df$c, TRUE)
## End(Not run)

[Package SparkR version 2.1.2 Index]