from_utc_timestamp {SparkR}	R Documentation

from_utc_timestamp

Description

Given a timestamp, which corresponds to a certain time of day in UTC, returns another timestamp that corresponds to the same time of day in the given timezone.

Usage

## S4 method for signature 'Column,character'
from_utc_timestamp(y, x)

from_utc_timestamp(y, x)

Arguments

y
Column to compute on.

x
time zone to use.

Note

from_utc_timestamp since 1.5.0

See Also

Other datetime_funcs: add_months, add_months, add_months,Column,numeric-method; date_add, date_add, date_add,Column,numeric-method; date_format, date_format, date_format,Column,character-method; date_sub, date_sub, date_sub,Column,numeric-method; datediff, datediff, datediff,Column-method; dayofmonth, dayofmonth, dayofmonth,Column-method; dayofyear, dayofyear, dayofyear,Column-method; from_unixtime, from_unixtime, from_unixtime,Column-method; hour, hour, hour,Column-method; last_day, last_day, last_day,Column-method; minute, minute, minute,Column-method; months_between, months_between, months_between,Column-method; month, month, month,Column-method; next_day, next_day, next_day,Column,character-method; quarter, quarter, quarter,Column-method; second, second, second,Column-method; to_date, to_date, to_date, to_date,Column,character-method, to_date,Column,missing-method; to_timestamp, to_timestamp, to_timestamp, to_timestamp,Column,character-method, to_timestamp,Column,missing-method; to_utc_timestamp, to_utc_timestamp, to_utc_timestamp,Column,character-method; unix_timestamp, unix_timestamp, unix_timestamp, unix_timestamp, unix_timestamp,Column,character-method, unix_timestamp,Column,missing-method, unix_timestamp,missing,missing-method; weekofyear, weekofyear, weekofyear,Column-method; window, window, window,Column-method; year, year, year,Column-method

Examples

## Not run: from_utc_timestamp(df$t, 'PST')

[Package SparkR version 2.2.0 Index]