generateAliasesForIntersectedCols {SparkR}	R Documentation

Creates a list of columns by replacing the intersected ones with aliases

Description

Creates a list of columns by replacing the intersected ones with aliases. The name of the alias column is formed by concatanating the original column name and a suffix.

Usage

generateAliasesForIntersectedCols(x, intersectedColNames, suffix)

Arguments

x
a SparkDataFrame

intersectedColNames
a list of intersected column names of the SparkDataFrame

suffix
a suffix for the column name

Value

list of columns

Note

generateAliasesForIntersectedCols since 1.6.0

[Package SparkR version 2.2.0 Index]