getLocalProperty {SparkR}	R Documentation

Get a local property set in this thread, or NULL if it is missing. See setLocalProperty.

Description

Get a local property set in this thread, or NULL if it is missing. See setLocalProperty.

Usage

getLocalProperty(key)

Arguments

key
The key for a local property.

Note

getLocalProperty since 2.3.0

Examples

## Not run: 
##D getLocalProperty("spark.scheduler.pool")
## End(Not run)

[Package SparkR version 3.2.3 Index]