greatest {SparkR}	R Documentation

greatest

Description

Returns the greatest value of the list of column names, skipping null values. This function takes at least 2 parameters. It will return null if all parameters are null.

Usage

## S4 method for signature 'Column'
greatest(x, ...)

greatest(x, ...)

See Also

Other normal_funcs: abs; bitwiseNOT, bitwiseNOT; col, column, column; expr, expr; ifelse; is.nan, isnan, isnan; least, least; lit, lit; nanvl, nanvl; negate, negate; randn, randn, randn; rand, rand, rand; struct, struct; when

Examples

## Not run: greatest(df$c, df$d)

[Package SparkR version 1.6.0 Index]