groupBy,DataFrame-method {SparkR}	R Documentation

GroupBy

Description

Groups the DataFrame using the specified columns, so we can run aggregation on them.

Usage

## S4 method for signature 'DataFrame'
groupBy(x, ...)

## S4 method for signature 'DataFrame'
group_by(x, ...)

group_by(x, ...)

Arguments

x
a DataFrame

Value

a GroupedData

See Also

GroupedData

Examples

## Not run: 
  # Compute the average for all numeric columns grouped by department.
  avg(groupBy(df, "department"))

  # Compute the max age and average salary, grouped by department and gender.
  agg(groupBy(df, "department", "gender"), salary="avg", "age" -> "max")

## End(Not run)

[Package SparkR version 1.4.0 Index]