ifelse {SparkR}	R Documentation

ifelse

Description

Evaluates a list of conditions and returns yes if the conditions are satisfied. Otherwise no is returned for unmatched conditions.

Usage

## S4 method for signature 'Column'
ifelse(test, yes, no)

Arguments

test
a Column expression that describes the condition.

yes
return values for TRUE elements of test.

no
return values for FALSE elements of test.

Note

ifelse since 1.5.0

See Also

when

Other normal_funcs: abs, bitwiseNOT, coalesce, column, expr, greatest, isnan, least, lit, nanvl, negate, randn, rand, struct, when

Examples

## Not run: 
##D ifelse(df$a > 1 & df$b > 2, 0, 1)
##D ifelse(df$a > 1, df$a, 1)
## End(Not run)

[Package SparkR version 2.1.2 Index]