invoke_higher_order_function {SparkR}	R Documentation

Invokes higher order function expression identified by name, (relative to o.a.s.sql.catalyst.expressions)

Description

Invokes higher order function expression identified by name, (relative to o.a.s.sql.catalyst.expressions)

Usage

invoke_higher_order_function(name, cols, funs)

Arguments

name
character

cols
list of character or Column objects

funs
list of named list(fun = ..., expected_narg = ...)

Value

a Column representing name applied to cols with funs

[Package SparkR version 3.2.3 Index]