least {SparkR}	R Documentation

least

Description

Returns the least value of the list of column names, skipping null values. This function takes at least 2 parameters. It will return null if all parameters are null.

Usage

## S4 method for signature 'Column'
least(x, ...)

least(x, ...)

Arguments

x
Column to compute on

...
other columns

Note

least since 1.5.0

See Also

Other normal_funcs: Column-class, column, column, column, column,character-method, column,jobj-method; abs, abs,Column-method; bitwiseNOT, bitwiseNOT, bitwiseNOT,Column-method; expr, expr, expr,character-method; greatest, greatest, greatest,Column-method; ifelse, ifelse,Column-method; is.nan, is.nan,Column-method, isnan, isnan, isnan,Column-method; lit, lit, lit,ANY-method; nanvl, nanvl, nanvl,Column-method; negate, negate, negate,Column-method; randn, randn, randn, randn,missing-method, randn,numeric-method; rand, rand, rand, rand,missing-method, rand,numeric-method; struct, struct, struct,characterOrColumn-method; when, when, when,Column-method

Examples

## Not run: least(df$c, df$d)

[Package SparkR version 2.0.1 Index]