listColumns {SparkR}	R Documentation

Returns a list of columns for the given table/view in the specified database

Description

Returns a list of columns for the given table/view in the specified database.

Usage

listColumns(tableName, databaseName = NULL)

Arguments

tableName
the qualified or unqualified name that designates a table/view. If no database identifier is provided, it refers to a table/view in the current database. If databaseName parameter is specified, this must be an unqualified name.

databaseName
(optional) name of the database

Value

a SparkDataFrame of the list of column descriptions.

Note

since 2.2.0

Examples

## Not run: 
##D sparkR.session()
##D listColumns("mytable")
## End(Not run)

[Package SparkR version 2.4.1 Index]