listFunctions {SparkR}	R Documentation

Returns a list of functions registered in the specified database

Description

Returns a list of functions registered in the specified database. This includes all temporary functions.

Usage

listFunctions(databaseName = NULL)

Arguments

databaseName
(optional) name of the database

Value

a SparkDataFrame of the list of function descriptions.

Note

since 2.2.0

Examples

## Not run: 
sparkR.session()
listFunctions()

## End(Not run)

[Package SparkR version 2.2.0 Index]