listTables {SparkR}	R Documentation

Returns a list of tables or views in the specified database

Description

Returns a list of tables or views in the specified database. This includes all temporary views.

Usage

listTables(databaseName = NULL)

Arguments

databaseName
(optional) name of the database

Value

a SparkDataFrame of the list of tables.

Note

since 2.2.0

See Also

tables

Examples

## Not run: 
sparkR.session()
listTables()
listTables("default")

## End(Not run)

[Package SparkR version 2.2.0 Index]