locate {SparkR}	R Documentation

locate

Description

Locate the position of the first occurrence of substr.

Usage

locate(substr, str, ...)

## S4 method for signature 'character,Column'
locate(substr, str, pos = 1)

Arguments

substr
a character string to be matched.

str
a Column where matches are sought for each entry.

...
further arguments to be passed to or from other methods.

pos
start position of search.

Details

Note: The position is not zero based, but 1 based index. Returns 0 if substr could not be found in str.

Note

locate since 1.5.0

See Also

Other string_funcs: ascii, base64, concat_ws, concat, decode, encode, format_number, format_string, initcap, instr, length, levenshtein, lower, lpad, ltrim, regexp_extract, regexp_replace, reverse, rpad, rtrim, soundex, substring_index, translate, trim, unbase64, upper

Examples

## Not run: locate('b', df$c, 1)

[Package SparkR version 2.1.2 Index]