dropna {SparkR}	R Documentation

dropna

Description

Returns a new DataFrame omitting rows with null values.

Replace null values.

Usage

## S4 method for signature 'DataFrame'
dropna(x, how = c("any", "all"), minNonNulls = NULL,
  cols = NULL)

## S4 method for signature 'DataFrame'
na.omit(object, how = c("any", "all"),
  minNonNulls = NULL, cols = NULL)

## S4 method for signature 'DataFrame'
fillna(x, value, cols = NULL)

dropna(x, how = c("any", "all"), minNonNulls = NULL, cols = NULL)

na.omit(object, ...)

fillna(x, value, cols = NULL)

Arguments

x
A SparkSQL DataFrame.

how
"any" or "all". if "any", drop a row if it contains any nulls. if "all", drop a row only if all its values are null. if minNonNulls is specified, how is ignored.

minNonNulls
If specified, drop rows that have less than minNonNulls non-null values. This overwrites the how parameter.

cols
Optional list of column names to consider.

value
Value to replace null values with. Should be an integer, numeric, character or named list. If the value is a named list, then cols is ignored and value must be a mapping from column name (character) to replacement value. The replacement value must be an integer, numeric or character.

x
A SparkSQL DataFrame.

cols
optional list of column names to consider. Columns specified in cols that do not have matching data type are ignored. For example, if value is a character, and subset contains a non-character column, then the non-character column is simply ignored.

Value

A DataFrame

A DataFrame

Examples

## Not run: 
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlCtx, path)
dropna(df)

## End(Not run)
## Not run: 
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlCtx, path)
fillna(df, 1)
fillna(df, list("age" = 20, "name" = "unknown"))

## End(Not run)

[Package SparkR version 1.5.1 Index]