objectFile {SparkR}	R Documentation

Load an RDD saved as a SequenceFile containing serialized objects.

Description

The file to be loaded should be one that was previously generated by calling saveAsObjectFile() of the RDD class.

Usage

objectFile(sc, path, minSplits = NULL)

Arguments

sc
SparkContext to use

path
Path of file to read. A vector of multiple paths is allowed.

minSplits
Minimum number of splits to be created. If NULL, the default value is chosen based on available parallelism.

Value

RDD containing serialized R objects.

See Also

saveAsObjectFile

Examples

## Not run: 
##D  sc <- sparkR.init()
##D  rdd <- objectFile(sc, "myfile")
## End(Not run)

[Package SparkR version 0.1 Index]