pipeRDD {SparkR}	R Documentation

Pipes elements to a forked external process.

Description

The same as 'pipe()' in Spark.

Usage

pipeRDD(rdd, command, env = list())

## S4 method for signature 'RDD,character'
pipeRDD(rdd, command, env = list())

Arguments

rdd
The RDD whose elements are piped to the forked external process.

command
The command to fork an external process.

env
A named list to set environment variables of the external process.

Value

A new RDD created by piping all elements to a forked external process.

Examples

## Not run: 
##D sc <- sparkR.init()
##D rdd <- parallelize(sc, 1:10)
##D collect(pipeRDD(rdd, "more")
##D Output: c("1", "2", ..., "10")
## End(Not run)

[Package SparkR version 0.1 Index]