rand {SparkR}	R Documentation

rand

Description

Generate a random column with independent and identically distributed (i.i.d.) samples from U[0.0, 1.0].

Usage

rand(seed)

## S4 method for signature 'missing'
rand(seed)

## S4 method for signature 'numeric'
rand(seed)

Arguments

seed
a random seed. Can be missing.

Note

rand since 1.5.0

rand(numeric) since 1.5.0

See Also

Other normal_funcs: abs, bitwiseNOT, coalesce, column, expr, greatest, ifelse, isnan, least, lit, nanvl, negate, randn, struct, when

Examples

## Not run: rand()

[Package SparkR version 2.1.2 Index]