randn {SparkR}	R Documentation

randn

Description

Generate a column with independent and identically distributed (i.i.d.) samples from the standard normal distribution.

Usage

randn(seed)

## S4 method for signature 'missing'
randn(seed)

## S4 method for signature 'numeric'
randn(seed)

Arguments

seed
a random seed. Can be missing.

Note

randn since 1.5.0

randn(numeric) since 1.5.0

See Also

Other normal_funcs: abs, bitwiseNOT, coalesce, column, expr, greatest, ifelse, isnan, least, lit, nanvl, negate, rand, struct, when

Examples

## Not run: randn()

[Package SparkR version 2.1.2 Index]