rank {SparkR}	R Documentation

rank

Description

Window function: returns the rank of rows within a window partition.

Usage

rank(x, ...)

## S4 method for signature 'missing'
rank()

## S4 method for signature 'ANY'
rank(x, ...)

Arguments

x
a numeric, complex, character or logical vector.

...
additional argument(s) passed to the method.

Details

The difference between rank and dense_rank is that dense_rank leaves no gaps in ranking sequence when there are ties. That is, if you were ranking a competition using dense_rank and had three people tie for second place, you would say that all three were in second place and that the next person came in third. Rank would give me sequential numbers, making the person that came in third place (after the ties) would register as coming in fifth.

This is equivalent to the RANK function in SQL.

Note

rank since 1.6.0

See Also

Other window_funcs: cume_dist, dense_rank, lag, lead, ntile, percent_rank, row_number

Examples

## Not run: 
##D   df <- createDataFrame(mtcars)
##D   ws <- orderBy(windowPartitionBy("am"), "hp")
##D   out <- select(df, over(rank(), ws), df$hp, df$am)
## End(Not run)

[Package SparkR version 2.1.2 Index]