read.stream {SparkR}	R Documentation

Load a streaming SparkDataFrame

Description

Returns the dataset in a data source as a SparkDataFrame

Usage

read.stream(source = NULL, schema = NULL, ...)

Arguments

source
The name of external data source

schema
The data schema defined in structType, this is required for file-based streaming data source

...
additional external data source specific named options, for instance path for file-based streaming data source

Details

The data source is specified by the source and a set of options(...). If source is not specified, the default data source configured by "spark.sql.sources.default" will be used.

Value

SparkDataFrame

Note

read.stream since 2.2.0

experimental

See Also

write.stream

Examples

## Not run: 
sparkR.session()
df <- read.stream("socket", host = "localhost", port = 9999)
q <- write.stream(df, "text", path = "/home/user/out", checkpointLocation = "/home/user/cp")

df <- read.stream("json", path = jsonDir, schema = schema, maxFilesPerTrigger = 1)

## End(Not run)

[Package SparkR version 2.2.0 Index]