read.text {SparkR}	R Documentation

Create a SparkDataFrame from a text file.

Description

Loads text files and returns a SparkDataFrame whose schema starts with a string column named "value", and followed by partitioned columns if there are any. The text files must be encoded as UTF-8.

Usage

read.text(path, ...)

Arguments

path
Path of file to read. A vector of multiple paths is allowed.

...
additional external data source specific named properties. You can find the text-specific options for reading text files in https://spark.apache.org/docs/latest/sql-data-sources-text.html#data-source-option Data Source Option in the version you use.

Details

Each line in the text file is a new row in the resulting SparkDataFrame.

Value

SparkDataFrame

Note

read.text since 1.6.1

Examples

## Not run: 
##D sparkR.session()
##D path <- "path/to/file.txt"
##D df <- read.text(path)
## End(Not run)

[Package SparkR version 3.2.2 Index]