recoverPartitions {SparkR}	R Documentation

Recovers all the partitions in the directory of a table and update the catalog

Description

Recovers all the partitions in the directory of a table and update the catalog. The name should reference a partitioned table, and not a view.

Usage

recoverPartitions(tableName)

Arguments

tableName
the qualified or unqualified name that designates a table. If no database identifier is provided, it refers to a table in the current database.

Note

since 2.2.0

Examples

## Not run: 
##D sparkR.session()
##D recoverPartitions("myTable")
## End(Not run)

[Package SparkR version 2.4.1 Index]