refreshTable {SparkR}	R Documentation

Invalidates and refreshes all the cached data and metadata of the given table

Description

Invalidates and refreshes all the cached data and metadata of the given table. For performance reasons, Spark SQL or the external data source library it uses might cache certain metadata about a table, such as the location of blocks. When those change outside of Spark SQL, users should call this function to invalidate the cache.

Usage

refreshTable(tableName)

Arguments

tableName
the qualified or unqualified name that designates a table. If no database identifier is provided, it refers to a table in the current database.

Details

If this table is cached as an InMemoryRelation, drop the original cached version and make the new version cached lazily.

Note

since 2.2.0

Examples

## Not run: 
##D sparkR.session()
##D refreshTable("myTable")
## End(Not run)

[Package SparkR version 3.2.2 Index]