saveAsTextFile {SparkR}	R Documentation

Save this RDD as a text file, using string representations of elements.

Description

Save this RDD as a text file, using string representations of elements.

Usage

saveAsTextFile(rdd, path)

## S4 method for signature 'RDD,character'
saveAsTextFile(rdd, path)

Arguments

rdd
The RDD to save

path
The directory where the splits of the text file are saved

Examples

## Not run: 
##D sc <- sparkR.init()
##D rdd <- parallelize(sc, 1:3)
##D saveAsTextFile(rdd, "/tmp/sparkR-tmp")
## End(Not run)

[Package SparkR version 0.1 Index]