showDF {SparkR}	R Documentation

showDF

Description

Print the first numRows rows of a SparkDataFrame

Usage

showDF(x, ...)

## S4 method for signature 'SparkDataFrame'
showDF(x, numRows = 20, truncate = TRUE, vertical = FALSE)

Arguments

x
a SparkDataFrame.

...
further arguments to be passed to or from other methods.

numRows
the number of rows to print. Defaults to 20.

truncate
whether truncate long strings. If TRUE, strings more than 20 characters will be truncated. However, if set greater than zero, truncates strings longer than truncate characters and all cells will be aligned right.

vertical
whether print output rows vertically (one line per column value).

Note

showDF since 1.4.0

See Also

Other SparkDataFrame functions: SparkDataFrame-class, agg(), alias(), arrange(), as.data.frame(), attach,SparkDataFrame-method, broadcast(), cache(), checkpoint(), coalesce(), collect(), colnames(), coltypes(), createOrReplaceTempView(), crossJoin(), cube(), dapplyCollect(), dapply(), describe(), dim(), distinct(), dropDuplicates(), dropna(), drop(), dtypes(), exceptAll(), except(), explain(), filter(), first(), gapplyCollect(), gapply(), getNumPartitions(), group_by(), head(), hint(), histogram(), insertInto(), intersectAll(), intersect(), isLocal(), isStreaming(), join(), limit(), localCheckpoint(), merge(), mutate(), ncol(), nrow(), persist(), printSchema(), randomSplit(), rbind(), rename(), repartitionByRange(), repartition(), rollup(), sample(), saveAsTable(), schema(), selectExpr(), select(), show(), storageLevel(), str(), subset(), summary(), take(), toJSON(), unionAll(), unionByName(), union(), unpersist(), withColumn(), withWatermark(), with(), write.df(), write.jdbc(), write.json(), write.orc(), write.parquet(), write.stream(), write.text()

Examples

## Not run: 
##D sparkR.session()
##D path <- "path/to/file.json"
##D df <- read.json(path)
##D showDF(df)
## End(Not run)

[Package SparkR version 3.2.2 Index]