sparkR.init {SparkR}	R Documentation

Initialize a new Spark Context.

Description

This function initializes a new SparkContext.

Usage

sparkR.init(master = "", appName = "SparkR",
  sparkHome = Sys.getenv("SPARK_HOME"), sparkEnvir = list(),
  sparkExecutorEnv = list(), sparkJars = "", sparkPackages = "")

Arguments

master
The Spark master URL.

appName
Application name to register with cluster manager

sparkHome
Spark Home directory

sparkEnvir
Named list of environment variables to set on worker nodes.

sparkExecutorEnv
Named list of environment variables to be used when launching executors.

sparkJars
Character string vector of jar files to pass to the worker nodes.

sparkPackages
Character string vector of packages from spark-packages.org

Examples

## Not run: 
sc <- sparkR.init("local[2]", "SparkR", "/home/spark")
sc <- sparkR.init("local[2]", "SparkR", "/home/spark",
                 list(spark.executor.memory="1g"))
sc <- sparkR.init("yarn-client", "SparkR", "/home/spark",
                 list(spark.executor.memory="1g"),
                 list(LD_LIBRARY_PATH="/directory of JVM libraries (libjvm.so) on workers/"),
                 c("jarfile1.jar","jarfile2.jar"))

## End(Not run)

[Package SparkR version 1.5.1 Index]