sparkR.version {SparkR}	R Documentation

Get version of Spark on which this application is running

Description

Get version of Spark on which this application is running.

Usage

sparkR.version()

Value

a character string of the Spark version

Note

sparkR.version since 2.0.1

Examples

## Not run: 
##D sparkR.session()
##D version <- sparkR.version()
## End(Not run)

[Package SparkR version 2.4.1 Index]