sparkRSQL.init {SparkR}	R Documentation

(Deprecated) Initialize a new SQLContext

Description

This function creates a SparkContext from an existing JavaSparkContext and then uses it to initialize a new SQLContext

Usage

sparkRSQL.init(jsc = NULL)

Arguments

jsc
The existing JavaSparkContext created with SparkR.init()

Details

Starting SparkR 2.0, a SparkSession is initialized and returned instead. This API is deprecated and kept for backward compatibility only.

Note

sparkRSQL.init since 1.4.0

See Also

sparkR.session

Examples

## Not run: 
##D sc <- sparkR.init()
##D sqlContext <- sparkRSQL.init(sc)
## End(Not run)

[Package SparkR version 3.2.4 Index]