#' Give a random idea or "spark" of an activity to do.

#' Give a random idea or "spark" of an activity to do.

Usage

sparkr(give_idea = TRUE, category = NULL)

Arguments

give_idea

Logical. Set to TRUE to get a randomly selected idea of something to do.

category

An optional category to filter ideas (e.g., Physical, Wellness, Creative, Social, Bad).

Value

A random idea from a predefined list of activities or a message if no idea is selected.

Examples

sparkr(TRUE)              # Get a random idea
#> [1] "\"Appreciate everything around you\""
sparkr(give_idea = TRUE) # Get a random idea
#> [1] "\"Stick you hands in a wasp nest\""
sparkr(give_idea = FALSE) # Responds without providing an idea
#> [1] "Oh you came up with an idea all by yourself? Nice free will"