sparkr: Gives an idea or spark of inspiration for an activity to do

A package for giving you a random idea of something to do when you are lacking inspiration. Ideas are predetermined in a list and randomly selected for you to enjoy. Arguments for this function are simply TRUE or FALSE.

This package is an exercise in package creation using R studio. The package includes a sample function and a sample dataset with their respective documentation.

See also

Author

Maintainer: Liam Surry surryl@mcmaster.ca (ORCID)

Liam Surry, Department of Chemistry and Chemical Biology surryl@mcmaster.ca