startsWith {SparkR}	R Documentation

startsWith

Description

Determines if entries of x start with string (entries of) prefix respectively, where strings are recycled to common lengths.

Usage

startsWith(x, prefix)

## S4 method for signature 'Column'
startsWith(x, prefix)

Arguments

x
vector of character string whose "starts" are considered

prefix
character vector (often of length one)

Note

startsWith since 1.4.0

See Also

Other column_func: alias(), between(), cast(), endsWith(), otherwise(), over(), substr()

[Package SparkR version 3.2.2 Index]