struct {SparkR}	R Documentation

struct

Description

Creates a new struct column that composes multiple input columns.

Usage

struct(x, ...)

## S4 method for signature 'characterOrColumn'
struct(x, ...)

Arguments

x
a column to compute on.

...
optional column(s) to be included.

Note

struct since 1.6.0

See Also

Other normal_funcs: abs, bitwiseNOT, coalesce, column, expr, greatest, ifelse, isnan, least, lit, nanvl, negate, randn, rand, when

Examples

## Not run: 
##D struct(df$c, df$d)
##D struct("col1", "col2")
## End(Not run)

[Package SparkR version 2.1.2 Index]