structType {SparkR}	R Documentation

structType

Description

Create a structType object that contains the metadata for a SparkDataFrame. Intended for use with createDataFrame and toDF.

Usage

structType(x, ...)

## S3 method for class 'jobj'
structType(x, ...)

## S3 method for class 'structField'
structType(x, ...)

## S3 method for class 'character'
structType(x, ...)

Arguments

x
a structField object (created with the structField method). Since Spark 2.3, this can be a DDL-formatted string, which is a comma separated list of field definitions, e.g., "a INT, b STRING".

...
additional structField objects

Value

a structType object

Note

structType since 1.4.0

Examples

## Not run: 
##D schema <- structType(structField("a", "integer"), structField("c", "string"),
##D                       structField("avg", "double"))
##D df1 <- gapply(df, list("a", "c"),
##D               function(key, x) { y <- data.frame(key, mean(x$b), stringsAsFactors = FALSE) },
##D               schema)
##D schema <- structType("a INT, c STRING, avg DOUBLE")
##D df1 <- gapply(df, list("a", "c"),
##D               function(key, x) { y <- data.frame(key, mean(x$b), stringsAsFactors = FALSE) },
##D               schema)
## End(Not run)

[Package SparkR version 2.4.1 Index]