substring_index {SparkR}	R Documentation

substring_index

Description

Returns the substring from string str before count occurrences of the delimiter delim. If count is positive, everything the left of the final delimiter (counting from left) is returned. If count is negative, every to the right of the final delimiter (counting from the right) is returned. substring_index performs a case-sensitive match when searching for delim.

Usage

## S4 method for signature 'Column,character,numeric'
substring_index(x, delim, count)

substring_index(x, delim, count)

See Also

Other string_funcs: ascii, ascii; base64, base64; concat_ws, concat_ws; concat, concat; decode, decode; encode, encode; format_number, format_number; format_string, format_string; initcap, initcap; instr, instr; length; levenshtein, levenshtein; locate, locate; lower, lower; lpad, lpad; ltrim, ltrim; regexp_extract, regexp_extract; regexp_replace, regexp_replace; reverse, reverse; rpad, rpad; rtrim, rtrim; soundex, soundex; translate, translate; trim, trim; unbase64, unbase64; upper, upper

Examples

## Not run: 
##D substring_index(df$c, '.', 2)
##D substring_index(df$c, '.', -1)
## End(Not run)

[Package SparkR version 1.6.3 Index]