take,DataFrame,numeric-method {SparkR}	R Documentation

Take the first NUM rows of a DataFrame and return a the results as a data.frame

Description

Take the first NUM rows of a DataFrame and return a the results as a data.frame

Usage

## S4 method for signature 'DataFrame,numeric'
take(x, num)

Examples

## Not run: 
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
take(df, 2)

## End(Not run)

[Package SparkR version 1.4.0 Index]