windowPartitionBy {SparkR}	R Documentation

windowPartitionBy

Description

Creates a WindowSpec with the partitioning defined.

Usage

windowPartitionBy(col, ...)

## S4 method for signature 'character'
windowPartitionBy(col, ...)

## S4 method for signature 'Column'
windowPartitionBy(col, ...)

Arguments

col
A column name or Column by which rows are partitioned to windows.

...
Optional column names or Columns in addition to col, by which rows are partitioned to windows.

Note

windowPartitionBy(character) since 2.0.0

windowPartitionBy(Column) since 2.0.0

Examples

## Not run: 
  ws <- orderBy(windowPartitionBy("key1", "key2"), "key3")
  df1 <- select(df, over(lead("value", 1), ws))

  ws <- orderBy(windowPartitionBy(df$key1, df$key2), df$key3)
  df1 <- select(df, over(lead("value", 1), ws))

## End(Not run)

[Package SparkR version 2.2.0 Index]