with {SparkR}	R Documentation

Evaluate a R expression in an environment constructed from a SparkDataFrame

Description

Evaluate a R expression in an environment constructed from a SparkDataFrame with() allows access to columns of a SparkDataFrame by simply referring to their name. It appends every column of a SparkDataFrame into a new environment. Then, the given expression is evaluated in this new environment.

Usage

with(data, expr, ...)

## S4 method for signature 'SparkDataFrame'
with(data, expr, ...)

Arguments

data
(SparkDataFrame) SparkDataFrame to use for constructing an environment.

expr
(expression) Expression to evaluate.

...
arguments to be passed to future methods.

See Also

attach

Examples

## Not run: 
##D with(irisDf, nrow(Sepal_Width))
## End(Not run)

[Package SparkR version 2.0.0 Index]