DataFrame-class {SparkR}	R Documentation

S4 class that represents a DataFrame

Description

DataFrames can be created using functions like jsonFile, table etc.

Usage

dataFrame(sdf, isCached = FALSE)

toRDD(x)

groupedData(sgd)

Arguments

sdf
A Java object reference to the backing Scala DataFrame

isCached
TRUE if the dataFrame is cached

Slots

env

An R environment that stores bookkeeping states of the DataFrame

sdf

A Java object reference to the backing Scala DataFrame

See Also

jsonFile, table

[Package SparkR version 1.5.1 Index]