approxCountDistinct {SparkR}	R Documentation

approxCountDistinct

Description

Aggregate function: returns the approximate number of distinct items in a group.

Approx Count Distinct

Usage

## S4 method for signature 'Column'
approxCountDistinct(x, rsd = 0.95)

## S4 method for signature 'Column'
approxCountDistinct(x, rsd = 0.95)

approxCountDistinct(x, ...)

Value

the approximate number of distinct items in a group.

See Also

Other agg_funcs: avg, avg; countDistinct, countDistinct, countDistinct, n_distinct; count, count, count, n, nrow; first, first; last, last; max; mean; min; sumDistinct, sumDistinct; sum

Other agg_funcs: avg, avg; countDistinct, countDistinct, countDistinct, n_distinct; count, count, count, n, nrow; first, first; last, last; max; mean; min; sumDistinct, sumDistinct; sum

Examples

## Not run: approxCountDistinct(df$c)

[Package SparkR version 1.5.2 Index]