Column-class {SparkR}	R Documentation

S4 class that represents a DataFrame column

Description

The column class supports unary, binary operations on DataFrame columns

Set a new name for a column

An expression that returns a substring.

Casts the column to a different data type.

Approx Count Distinct

Count Distinct

Usage

## S4 method for signature 'Column'
alias(object, data)

## S4 method for signature 'Column'
substr(x, start, stop)

## S4 method for signature 'Column'
cast(x, dataType)

## S4 method for signature 'Column'
approxCountDistinct(x, rsd = 0.95)

## S4 method for signature 'Column'
countDistinct(x, ...)

## S4 method for signature 'Column'
n_distinct(x, ...)

## S4 method for signature 'Column'
n(x)

approxCountDistinct(x, ...)

asc(x)

avg(x, ...)

cast(x, dataType)

cbrt(x)

contains(x, ...)

countDistinct(x, ...)

desc(x)

endsWith(x, ...)

getField(x, ...)

getItem(x, ...)

hypot(y, x)

isNull(x)

isNotNull(x)

last(x)

like(x, ...)

lower(x)

n(x)

n_distinct(x, ...)

rint(x, ...)

rlike(x, ...)

startsWith(x, ...)

sumDistinct(x)

toDegrees(x)

toRadians(x)

upper(x)

Arguments

start
starting position

stop
ending position

jc
reference to JVM DataFrame column

Value

the approximate number of distinct items in a group.

the number of distinct items in a group.

Examples

## Not run: 
##D   cast(df$age, "string")
##D   cast(df$name, list(type="array", elementType="byte", containsNull = TRUE))
## End(Not run)

[Package SparkR version 1.4.0 Index]