cume_dist {SparkR}	R Documentation

cume_dist

Description

Window function: returns the cumulative distribution of values within a window partition, i.e. the fraction of rows that are below the current row.

Usage

cume_dist(x = "missing")

## S4 method for signature 'missing'
cume_dist()

Arguments

x
empty. Should be used with no argument.

Details

N = total number of rows in the partition cume_dist(x) = number of values before (and including) x / N

This is equivalent to the CUME_DIST function in SQL.

Note

cume_dist since 1.6.0

See Also

Other window_funcs: dense_rank, lag, lead, ntile, percent_rank, rank, row_number

Examples

## Not run: 
##D   df <- createDataFrame(mtcars)
##D   ws <- orderBy(windowPartitionBy("am"), "hp")
##D   out <- select(df, over(cume_dist(), ws), df$hp, df$am)
## End(Not run)

[Package SparkR version 2.1.2 Index]