cume_dist {SparkR}	R Documentation

cume_dist

Description

Window function: returns the cumulative distribution of values within a window partition, i.e. the fraction of rows that are below the current row.

Usage

## S4 method for signature 'missing'
cume_dist()

cume_dist(x)

Details

N = total number of rows in the partition cume_dist(x) = number of values before (and including) x / N

This is equivalent to the CUME_DIST function in SQL.

See Also

Other window_funcs: dense_rank, dense_rank; lag, lag; lead, lead; ntile, ntile; percent_rank, percent_rank; rank, rank; row_number, row_number

Examples

## Not run: cume_dist()

[Package SparkR version 1.6.0 Index]