endsWith {SparkR}	R Documentation

endsWith

Description

Determines if entries of x end with string (entries of) suffix respectively, where strings are recycled to common lengths.

Usage

endsWith(x, suffix)

## S4 method for signature 'Column'
endsWith(x, suffix)

Arguments

x
vector of character string whose "ends" are considered

suffix
character vector (often of length one)

Note

endsWith since 1.4.0

See Also

Other column_func: alias(), between(), cast(), otherwise(), over(), startsWith(), substr()

[Package SparkR version 3.2.1 Index]