endsWith {SparkR}	R Documentation

endsWith

Description

Determines if entries of x end with string (entries of) suffix respectively, where strings are recycled to common lengths.

Usage

## S4 method for signature 'Column'
endsWith(x, suffix)

endsWith(x, suffix)

Arguments

x
vector of character string whose "ends" are considered

suffix
character vector (often of length one)

Note

endsWith since 1.4.0

See Also

Other colum_func: alias, alias,Column-method; between, between, between,Column-method; cast, cast, cast,Column-method; otherwise, otherwise, otherwise,Column-method; over, over, over,Column,WindowSpec-method; startsWith, startsWith, startsWith,Column-method; substr, substr,Column-method

[Package SparkR version 2.0.0 Index]