format_number {SparkR}	R Documentation

format_number

Description

Formats numeric column x to a format like '#,###,###.##', rounded to d decimal places, and returns the result as a string column.

Usage

## S4 method for signature 'Column,numeric'
format_number(y, x)

format_number(y, x)

Details

If d is 0, the result has no decimal point or fractional part. If d < 0, the result will be null.'

See Also

Other string_funcs: ascii, ascii; base64, base64; concat_ws, concat_ws; concat, concat; format_string, format_string; initcap, initcap; instr, instr; length; levenshtein, levenshtein; locate, locate; lower, lower; lpad, lpad; ltrim, ltrim; regexp_extract, regexp_extract; regexp_replace, regexp_replace; reverse, reverse; rpad, rpad; rtrim, rtrim; soundex, soundex; substring_index, substring_index; translate, translate; trim, trim; unbase64, unbase64; upper, upper

[Package SparkR version 1.5.1 Index]