from_unixtime {SparkR}	R Documentation

from_unixtime

Description

Converts the number of seconds from unix epoch (1970-01-01 00:00:00 UTC) to a string representing the timestamp of that moment in the current system time zone in the given format.

Usage

## S4 method for signature 'Column'
from_unixtime(x, format = "yyyy-MM-dd HH:mm:ss")

from_unixtime(x, ...)

See Also

Other datetime_funcs: add_months, add_months; date_add, date_add; date_format, date_format; date_sub, date_sub; datediff, datediff; dayofmonth, dayofmonth; dayofyear, dayofyear; from_utc_timestamp, from_utc_timestamp; hour, hour; last_day, last_day; minute, minute; months_between, months_between; month, month; next_day, next_day; quarter, quarter; second, second; to_date, to_date; to_utc_timestamp, to_utc_timestamp; unix_timestamp, unix_timestamp, unix_timestamp, unix_timestamp; weekofyear, weekofyear; year, year

[Package SparkR version 1.5.0 Index]