from_utc_timestamp {SparkR}	R Documentation

from_utc_timestamp

Description

Given a timestamp, which corresponds to a certain time of day in UTC, returns another timestamp that corresponds to the same time of day in the given timezone.

Usage

from_utc_timestamp(y, x)

## S4 method for signature 'Column,character'
from_utc_timestamp(y, x)

Arguments

y
Column to compute on.

x
time zone to use.

Note

from_utc_timestamp since 1.5.0

See Also

Other datetime_funcs: add_months, date_add, date_format, date_sub, datediff, dayofmonth, dayofyear, from_unixtime, hour, last_day, minute, months_between, month, next_day, quarter, second, to_date, to_utc_timestamp, unix_timestamp, weekofyear, window, year

Examples

## Not run: from_utc_timestamp(df$t, 'PST')

[Package SparkR version 2.1.2 Index]