groupBy {SparkR}	R Documentation

GroupBy

Description

Groups the DataFrame using the specified columns, so we can run aggregation on them.

Usage

## S4 method for signature 'DataFrame'
groupBy(x, ...)

## S4 method for signature 'DataFrame'
group_by(x, ...)

group_by(x, ...)

groupBy(x, ...)

Arguments

x
a DataFrame

Value

a GroupedData

See Also

GroupedData

Other DataFrame functions: $, $<-, select, select, select,DataFrame,Column-method, select,DataFrame,list-method, selectExpr; DataFrame-class, dataFrame, groupedData; [, [, [[, subset; agg, agg, count,GroupedData-method, summarize, summarize; arrange, arrange, arrange, orderBy, orderBy; as.data.frame, as.data.frame,DataFrame-method; attach, attach,DataFrame-method; cache; collect; colnames, colnames, colnames<-, colnames<-, columns, names, names<-; coltypes, coltypes, coltypes<-, coltypes<-; columns, dtypes, printSchema, schema, schema; count, nrow; describe, describe, describe, summary, summary, summary,PipelineModel-method; dim; distinct, unique; dropna, dropna, fillna, fillna, na.omit, na.omit; dtypes; except, except; explain, explain; filter, filter, where, where; first, first; head; insertInto, insertInto; intersect, intersect; isLocal, isLocal; join; limit, limit; merge, merge; mutate, mutate, transform; ncol; persist; printSchema; rbind, rbind, unionAll, unionAll; registerTempTable, registerTempTable; rename, rename, withColumnRenamed, withColumnRenamed; repartition; sample, sample, sample_frac, sample_frac; saveAsParquetFile, saveAsParquetFile, write.parquet, write.parquet; saveAsTable, saveAsTable; saveDF, saveDF, write.df, write.df; selectExpr; showDF, showDF; show, show, show,GroupedData-method; take; transform, withColumn, withColumn; unpersist; write.json, write.json

Examples

## Not run: 
##D   # Compute the average for all numeric columns grouped by department.
##D   avg(groupBy(df, "department"))
##D 
##D   # Compute the max age and average salary, grouped by department and gender.
##D   agg(groupBy(df, "department", "gender"), salary="avg", "age" -> "max")
## End(Not run)

[Package SparkR version 1.6.0 Index]