join {SparkR}	R Documentation

Join

Description

Joins two SparkDataFrames based on the given join expression.

Usage

## S4 method for signature 'SparkDataFrame,SparkDataFrame'
join(x, y, joinExpr = NULL,
  joinType = NULL)

Arguments

x
A SparkDataFrame

y
A SparkDataFrame

joinExpr
(Optional) The expression used to perform the join. joinExpr must be a Column expression. If joinExpr is omitted, the default, inner join is attempted and an error is thrown if it would be a Cartesian Product. For Cartesian join, use crossJoin instead.

joinType
The type of join to perform, default 'inner'. Must be one of: 'inner', 'cross', 'outer', 'full', 'full_outer', 'left', 'left_outer', 'right', 'right_outer', 'left_semi', or 'left_anti'.

Value

A SparkDataFrame containing the result of the join operation.

Note

join since 1.4.0

See Also

merge crossJoin

Other SparkDataFrame functions: SparkDataFrame-class, agg, alias, arrange, as.data.frame, attach,SparkDataFrame-method, broadcast, cache, checkpoint, coalesce, collect, colnames, coltypes, createOrReplaceTempView, crossJoin, cube, dapplyCollect, dapply, describe, dim, distinct, dropDuplicates, dropna, drop, dtypes, exceptAll, except, explain, filter, first, gapplyCollect, gapply, getNumPartitions, group_by, head, hint, histogram, insertInto, intersectAll, intersect, isLocal, isStreaming, limit, localCheckpoint, merge, mutate, ncol, nrow, persist, printSchema, randomSplit, rbind, rename, repartitionByRange, repartition, rollup, sample, saveAsTable, schema, selectExpr, select, showDF, show, storageLevel, str, subset, summary, take, toJSON, unionByName, union, unpersist, withColumn, withWatermark, with, write.df, write.jdbc, write.json, write.orc, write.parquet, write.stream, write.text

Examples

## Not run: 
##D sparkR.session()
##D df1 <- read.json(path)
##D df2 <- read.json(path2)
##D join(df1, df2, df1$col1 == df2$col2) # Performs an inner join based on expression
##D join(df1, df2, df1$col1 == df2$col2, "right_outer")
##D join(df1, df2) # Attempts an inner join
## End(Not run)

[Package SparkR version 2.4.1 Index]