last {SparkR}	R Documentation

last

Description

Aggregate function: returns the last value in a group.

Usage

## S4 method for signature 'characterOrColumn'
last(x, na.rm = FALSE)

last(x, ...)

Arguments

x
column to compute on.

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

...
further arguments to be passed to or from other methods.

Details

The function by default returns the last values it sees. It will return the last non-missing value it sees when na.rm is set to true. If all values are missing, then NA is returned.

Note

last since 1.4.0

See Also

Other agg_funcs: agg, agg, agg, agg,GroupedData-method, agg,SparkDataFrame-method, summarize, summarize, summarize, summarize,GroupedData-method, summarize,SparkDataFrame-method; avg, avg, avg,Column-method; countDistinct, countDistinct, countDistinct,Column-method, n_distinct, n_distinct, n_distinct,Column-method; count, count, count,Column-method, count,GroupedData-method, n, n, n,Column-method; first, first, first, first,SparkDataFrame-method, first,characterOrColumn-method; kurtosis, kurtosis, kurtosis,Column-method; max, max,Column-method; mean, mean,Column-method; min, min,Column-method; sd, sd, sd,Column-method, stddev, stddev, stddev,Column-method; skewness, skewness, skewness,Column-method; stddev_pop, stddev_pop, stddev_pop,Column-method; stddev_samp, stddev_samp, stddev_samp,Column-method; sumDistinct, sumDistinct, sumDistinct,Column-method; sum, sum,Column-method; var_pop, var_pop, var_pop,Column-method; var_samp, var_samp, var_samp,Column-method; var, var, var,Column-method, variance, variance, variance,Column-method

Examples

## Not run: 
##D last(df$c)
##D last(df$c, TRUE)
## End(Not run)

[Package SparkR version 2.2.0 Index]