lit {SparkR}	R Documentation

lit

Description

A new Column is created to represent the literal value. If the parameter is a Column, it is returned unchanged.

Usage

lit(x)

## S4 method for signature 'ANY'
lit(x)

Arguments

x
a literal value or a Column.

Note

lit since 1.5.0

See Also

Other normal_funcs: abs, bitwiseNOT, coalesce, column, expr, greatest, ifelse, isnan, least, nanvl, negate, randn, rand, struct, when

Examples

## Not run: 
##D lit(df$name)
##D select(df, lit("x"))
##D select(df, lit("2015-01-01"))
## End(Not run)

[Package SparkR version 2.1.2 Index]