lit {SparkR}	R Documentation

lit

Description

A new Column is created to represent the literal value. If the parameter is a Column, it is returned unchanged.

Usage

## S4 method for signature 'ANY'
lit(x)

lit(x)

See Also

Other normal_funcs: abs; bitwiseNOT, bitwiseNOT; col, column, column; expr, expr; greatest, greatest; ifelse; is.nan, isnan, isnan; least, least; nanvl, nanvl; negate, negate; randn, randn, randn; rand, rand, rand; struct, struct; when

Examples

## Not run: 
##D lit(df$name)
##D select(df, lit("x"))
##D select(df, lit("2015-01-01"))
## End(Not run)

[Package SparkR version 1.6.1 Index]