dropna {SparkR}	R Documentation

dropna

Description

Returns a new SparkDataFrame omitting rows with null values.

Replace null values.

Usage

dropna(x, how = c("any", "all"), minNonNulls = NULL, cols = NULL)

na.omit(object, ...)

fillna(x, value, cols = NULL)

## S4 method for signature 'SparkDataFrame'
dropna(x, how = c("any", "all"),
  minNonNulls = NULL, cols = NULL)

## S4 method for signature 'SparkDataFrame'
na.omit(object, how = c("any", "all"),
  minNonNulls = NULL, cols = NULL)

## S4 method for signature 'SparkDataFrame'
fillna(x, value, cols = NULL)

Arguments

x
A SparkDataFrame.

how
"any" or "all". if "any", drop a row if it contains any nulls. if "all", drop a row only if all its values are null. if minNonNulls is specified, how is ignored.

minNonNulls
If specified, drop rows that have less than minNonNulls non-null values. This overwrites the how parameter.

cols
Optional list of column names to consider.

value
Value to replace null values with. Should be an integer, numeric, character or named list. If the value is a named list, then cols is ignored and value must be a mapping from column name (character) to replacement value. The replacement value must be an integer, numeric or character.

x
A SparkDataFrame.

cols
optional list of column names to consider. Columns specified in cols that do not have matching data type are ignored. For example, if value is a character, and subset contains a non-character column, then the non-character column is simply ignored.

Value

A SparkDataFrame

See Also

Other SparkDataFrame functions: SparkDataFrame-class, [[, agg, arrange, as.data.frame, attach, cache, collect, colnames, coltypes, columns, count, dapply, describe, dim, distinct, dropDuplicates, drop, dtypes, except, explain, filter, first, group_by, head, histogram, insertInto, intersect, isLocal, join, limit, merge, mutate, ncol, persist, printSchema, registerTempTable, rename, repartition, sample, saveAsTable, selectExpr, select, showDF, show, str, take, unionAll, unpersist, withColumn, write.df, write.jdbc, write.json, write.parquet, write.text

Examples

## Not run: 
##D sc <- sparkR.init()
##D sqlCtx <- sparkRSQL.init(sc)
##D path <- "path/to/file.json"
##D df <- read.json(sqlCtx, path)
##D dropna(df)
## End(Not run)
## Not run: 
##D sc <- sparkR.init()
##D sqlCtx <- sparkRSQL.init(sc)
##D path <- "path/to/file.json"
##D df <- read.json(sqlCtx, path)
##D fillna(df, 1)
##D fillna(df, list("age" = 20, "name" = "unknown"))
## End(Not run)

[Package SparkR version 2.0.0 Index]