read.df {SparkR}	R Documentation

Load an DataFrame

Description

Returns the dataset in a data source as a DataFrame

Usage

read.df(sqlContext, path = NULL, source = NULL, schema = NULL, ...)

Arguments

sqlContext
SQLContext to use

path
The path of files to load

source
the name of external data source

Details

The data source is specified by the 'source' and a set of options(...). If 'source' is not specified, the default data source configured by "spark.sql.sources.default" will be used.

Value

DataFrame

Examples

## Not run: 
##D sc <- sparkR.init()
##D sqlContext <- sparkRSQL.init(sc)
##D df <- read.df(sqlContext, "path/to/file.json", source = "json")
## End(Not run)

[Package SparkR version 1.5.0 Index]