read.stream {SparkR}	R Documentation

Load a streaming SparkDataFrame

Description

Returns the dataset in a data source as a SparkDataFrame

Usage

read.stream(source = NULL, schema = NULL, ...)

Arguments

source
The name of external data source

schema
The data schema defined in structType or a DDL-formatted string, this is required for file-based streaming data source

...
additional external data source specific named options, for instance path for file-based streaming data source. timeZone to indicate a timezone to be used to parse timestamps in the JSON/CSV data sources or partition values; If it isn't set, it uses the default value, session local timezone.

Details

The data source is specified by the source and a set of options(...). If source is not specified, the default data source configured by "spark.sql.sources.default" will be used.

Value

SparkDataFrame

Note

read.stream since 2.2.0

experimental

See Also

write.stream

Examples

## Not run: 
##D sparkR.session()
##D df <- read.stream("socket", host = "localhost", port = 9999)
##D q <- write.stream(df, "text", path = "/home/user/out", checkpointLocation = "/home/user/cp")
##D 
##D df <- read.stream("json", path = jsonDir, schema = schema, maxFilesPerTrigger = 1)
##D stringSchema <- "name STRING, info MAP<STRING, DOUBLE>"
##D df1 <- read.stream("json", path = jsonDir, schema = stringSchema, maxFilesPerTrigger = 1)
## End(Not run)

[Package SparkR version 2.4.1 Index]