rowsBetween {SparkR}	R Documentation

rowsBetween

Description

Defines the frame boundaries, from start (inclusive) to end (inclusive).

Usage

rowsBetween(x, start, end)

## S4 method for signature 'WindowSpec,numeric,numeric'
rowsBetween(x, start, end)

Arguments

x
a WindowSpec

start
boundary start, inclusive. The frame is unbounded if this is the minimum long value.

end
boundary end, inclusive. The frame is unbounded if this is the maximum long value.

Details

Both start and end are relative positions from the current row. For example, "0" means "current row", while "-1" means the row before the current row, and "5" means the fifth row after the current row.

Value

a WindowSpec

Note

rowsBetween since 2.0.0

See Also

Other windowspec_method: orderBy, partitionBy, rangeBetween

Examples

## Not run: 
##D   rowsBetween(ws, 0, 3)
## End(Not run)

[Package SparkR version 2.4.1 Index]