setJobDescription {SparkR}	R Documentation

Set a human readable description of the current job.

Description

Set a description that is shown as a job description in UI.

Usage

setJobDescription(value)

Arguments

value
The job description of the current job.

Note

setJobDescription since 2.3.0

Examples

## Not run: 
##D setJobDescription("This is an example job.")
## End(Not run)

[Package SparkR version 2.4.1 Index]