setJobGroup {SparkR}	R Documentation

Assigns a group ID to all the jobs started by this thread until the group ID is set to a different value or cleared.

Description

Assigns a group ID to all the jobs started by this thread until the group ID is set to a different value or cleared.

Usage

setJobGroup(sc, groupId, description, interruptOnCancel)

Arguments

sc
existing spark context

description
description for the the job group ID

interruptOnCancel
flag to indicate if the job is interrupted on job cancellation

groupid
the ID to be assigned to job groups

Examples

## Not run: 
##D sc <- sparkR.init()
##D setJobGroup(sc, "myJobGroup", "My job group description", TRUE)
## End(Not run)

[Package SparkR version 1.6.0 Index]