setLocalProperty {SparkR}	R Documentation

Set a local property that affects jobs submitted from this thread, such as the Spark fair scheduler pool.

Description

Set a local property that affects jobs submitted from this thread, such as the Spark fair scheduler pool.

Usage

setLocalProperty(key, value)

Arguments

key
The key for a local property.

value
The value for a local property.

Note

setLocalProperty since 2.3.0

Examples

## Not run: 
##D setLocalProperty("spark.scheduler.pool", "poolA")
## End(Not run)

[Package SparkR version 2.4.1 Index]