spark.getSparkFiles {SparkR}	R Documentation

Get the absolute path of a file added through spark.addFile.

Description

Get the absolute path of a file added through spark.addFile.

Usage

spark.getSparkFiles(fileName)

Arguments

fileName
The name of the file added through spark.addFile

Value

the absolute path of a file added through spark.addFile.

Note

spark.getSparkFiles since 2.1.0

Examples

## Not run: 
##D spark.getSparkFiles("myfile")
## End(Not run)

[Package SparkR version 2.4.1 Index]