spark.getSparkFilesRootDirectory {SparkR}	R Documentation

Get the root directory that contains files added through spark.addFile.

Description

Get the root directory that contains files added through spark.addFile.

Usage

spark.getSparkFilesRootDirectory()

Value

the root directory that contains files added through spark.addFile

Note

spark.getSparkFilesRootDirectory since 2.1.0

Examples

## Not run: 
##D spark.getSparkFilesRootDirectory()
## End(Not run)

[Package SparkR version 2.4.1 Index]