spark.survreg {SparkR}	R Documentation

Fit an accelerated failure time (AFT) survival regression model.

Description

Fit an accelerated failure time (AFT) survival regression model on a Spark DataFrame.

Usage

spark.survreg(data, formula, ...)

## S4 method for signature 'SparkDataFrame,formula'
spark.survreg(data, formula, ...)

Arguments

data
SparkDataFrame for training.

formula
A symbolic description of the model to be fitted. Currently only a few formula operators are supported, including '~', ':', '+', and '-'. Note that operator '.' is not supported currently.

Value

a fitted AFT survival regression model

See Also

survival: https://cran.r-project.org/web/packages/survival/

Examples

## Not run: 
##D df <- createDataFrame(sqlContext, ovarian)
##D model <- spark.survreg(df, Surv(futime, fustat) ~ ecog_ps + rx)
## End(Not run)

[Package SparkR version 2.0.0 Index]