startsWith {SparkR}	R Documentation

startsWith

Description

Determines if entries of x start with string (entries of) prefix respectively, where strings are recycled to common lengths.

Usage

## S4 method for signature 'Column'
startsWith(x, prefix)

startsWith(x, prefix)

Arguments

x
vector of character string whose "starts" are considered

prefix
character vector (often of length one)

Note

startsWith since 1.4.0

See Also

Other colum_func: alias, alias,Column-method; between, between, between,Column-method; cast, cast, cast,Column-method; endsWith, endsWith, endsWith,Column-method; otherwise, otherwise, otherwise,Column-method; over, over, over,Column,WindowSpec-method; substr, substr,Column-method

[Package SparkR version 2.0.0 Index]