[[ {SparkR}	R Documentation

Subset

Description

Return subsets of DataFrame according to given conditions

Usage

## S4 method for signature 'DataFrame,numericOrcharacter'
x[[i]]

## S4 method for signature 'DataFrame,missing,ANY'
x[i, j, ..., drop = TRUE]

## S4 method for signature 'DataFrame,Column,ANY'
x[i, j, ..., drop = TRUE]

## S4 method for signature 'DataFrame'
subset(x, subset, select, ...)

Arguments

x
A DataFrame

subset
A logical expression to filter on rows

select
expression for the single Column or a list of columns to select from the DataFrame

Value

A new DataFrame containing only the rows that meet the condition with selected columns

See Also

Other subsetting functions: $, $<-, select, select, select,DataFrame,Column-method, select,DataFrame,list-method, selectExpr; filter, filter, where, where

Examples

## Not run: 
##D   # Columns can be selected using `[[` and `[`
##D   df[[2]] == df[["age"]]
##D   df[,2] == df[,"age"]
##D   df[,c("name", "age")]
##D   # Or to filter rows
##D   df[df$age > 20,]
##D   # DataFrame can be subset on both rows and Columns
##D   df[df$name == "Smith", c(1,2)]
##D   df[df$age %in% c(19, 30), 1:2]
##D   subset(df, df$age %in% c(19, 30), 1:2)
##D   subset(df, df$age %in% c(19), select = c(1,2))
## End(Not run)

[Package SparkR version 1.5.0 Index]