unionAll {SparkR}	R Documentation

UnionAll

Description

Return a new DataFrame containing the union of rows in this DataFrame and another DataFrame. This is equivalent to 'UNION ALL' in SQL. Note that this does not remove duplicate rows across the two DataFrames.

Usage

## S4 method for signature 'DataFrame,DataFrame'
unionAll(x, y)

unionAll(x, y)

Arguments

x
A Spark DataFrame

y
A Spark DataFrame

Value

A DataFrame containing the result of the union.

Examples

## Not run: 
##D sc <- sparkR.init()
##D sqlContext <- sparkRSQL.init(sc)
##D df1 <- jsonFile(sqlContext, path)
##D df2 <- jsonFile(sqlContext, path2)
##D unioned <- unionAll(df, df2)
## End(Not run)

[Package SparkR version 1.5.2 Index]